object_cc.cc :  error in new long long int code for objectGet  .  int tmp -> long long int tmp
git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@824 baecf98e-1e81-4516-8920-d4c889e67ea4

+++ src/object_cc.cc
@@ -67,7 +67,7 @@ void objectGet(OBJECT* obj, const string& name, Long64& value, const string& def
 }
 void objectGet(OBJECT* obj, const string& name, long long int& value, const string& defVal)
 {
-   int tmp;
+   long long int  tmp;
    object_get(obj, name.c_str(), &tmp, NULL, 1, defVal.c_str());
    value = tmp;
 }
